% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scm_reg.R
\name{scm_reg}
\alias{scm_reg}
\title{Use to perform Stepwise Covariate Modeling (SCM)}
\usage{
scm_reg(
  dataset,
  variable,
  variable_event = NULL,
  weights_ordered = NULL,
  covariate.list = NULL,
  test_used = "Chisq",
  p_forward = 0.01,
  p_backward = 0.001,
  regression = "lm",
  cov_corr = NULL,
  search_direction = "full",
  base_relation = NULL,
  full_relation = NULL,
  max_steps = Inf,
  ...
)
}
\arguments{
\item{dataset}{name of dataset}

\item{variable}{dependent variable}

\item{variable_event}{event variable for coxph regression only}

\item{weights_ordered}{weight used in the ordered-categorical regression}

\item{covariate.list}{vector of covariate to be tested}

\item{test_used}{test to be applied. Can be 'Chisq' or 'AIC'}

\item{p_forward}{p value for forward selection and interaction phase}

\item{p_backward}{p value for forward step}

\item{regression}{type of regression, can be 'lm', 'logistic', 'coxph' or 'ordered-categorical'}

\item{cov_corr}{list of correlated covariates you don t want to be included simultaneously like eGFR and CRCL. Can be added as a list of vectors in case of multiple vectors of correlated covariates}

\item{search_direction}{step to be performed. Can be 'full', 'forward', 'interaction', 'backward', 'forward-backward', 'forward-interaction' or 'interaction-backward'}

\item{base_relation}{base relation included in the base model. Note those relationships cannot be removed in the backward step. Example "age*bwt + crcl"}

\item{full_relation}{starting model when performing backward step only.}

\item{max_steps}{maximal number of covariates included in the forward step}

\item{...}{Additional arguments}
}
\value{
Object of class \code{scmobject}
}
\description{
Use to perform Stepwise Covariate Modeling (SCM)
}
\examples{
 \dontrun{
 library(scmreg)
 library(MASS)
 haha <- scm_reg(dataset=housing,
                variable='Sat',
                covariate.list = c('Infl','Type','Cont'),
                p_forward=0.01,
                p_backward=0.001,
                test_used = 'AIC',
                regression='ordered-categorical',
                search_direction='forward-backward',
                weights_ordered='Freq',
                max_steps=Inf)

 }
}
