% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{tabscm}
\alias{tabscm}
\title{Table function for scm}
\usage{
tabscm(scmobject, ...)
}
\arguments{
\item{scmobject}{Object returned from \code{scm_reg}}

\item{...}{Additional args}
}
\value{
Object of class \code{flextable}
}
\description{
Table function for scm
}
