% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{theme_pps_table}
\alias{theme_pps_table}
\title{SCM Table Theme}
\usage{
theme_pps_table(x, font_name = "Times New Roman", font_size = 10)
}
\arguments{
\item{x}{Object of class \code{flextable}}

\item{font_name}{Character}

\item{font_size}{Numeric}
}
\value{

}
\description{
SCM Table Theme
}
